+++ src/App.vue
@@ -60,4 +60,20 @@
 #title
   font-size 24px
   font-weight 500
+
+.ripple
+  display block
+  position absolute
+  background-color rgba(32, 32, 32, 0.5)
+  border-radius 50%
+  transform scale(0)
+  z-index 1
+  animation ripple 0.4s linear
+
+@keyframes ripple {
+  to {
+    opacity 0
+    transform scale(2.5)
+  }
+}
 </style>
