+++ src/components/HomeView.vue
@@ -1,11 +1,6 @@
 <template>
   <div id="home">
-    <div class="create-container">
-      <div class="create-button">
-        <i class="material-icons">add</i>
-        Create New MindMap
-      </div>
-    </div>
+    <CreateButton />
     <div id="cards">
       <template v-for="item in testData">
         <Card :id="item.key" :key="item.key" :title="item.title" :body="item.body" />
@@ -18,10 +13,12 @@
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { mapState } from 'vuex';
 import Card from './Card.vue';
+import CreateButton from './CreateButton.vue';
 
 @Component({
   components: {
   Card,
+  CreateButton,
   },
   computed: {
   ...mapState([
@@ -44,40 +41,6 @@ export default class HomeView extends Vue {
   display grid
   grid-template-rows 52px 1fr
   grid-template-columns 1fr
-
-.create-container
-  /*--- position ---*/
-  grid-row 1 / 2
-  grid-column 1 / 2
-  /*--- end ---*/
-
-  /*--- layout ---*/
-  display flex
-  justify-content center
-  /*--- end ---*/
-
-.create-button
-  /*--- style ---*/
-  font-size 20px
-  font-weight 400
-  height 40px
-  width 260px
-  border-radius 4px
-  transition background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
-    box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
-    border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms
-  text-transform uppercase
-  /*--- end ---*/
-
-  /*--- layout ---*/
-  display flex
-  justify-content center
-  align-items center
-  /*--- end ---*/
-
-  &:hover
-    background-color rgba(255, 255, 255, 0.08)
-    cursor pointer
 
 #cards
   display grid
